Preserve fractional physical stackup sequence numbers
Amp-Thread-ID: https://ampcode.com/threads/T-01a07c28-80e5-72fd-9cc9-994ff3e716c3

+++ crates/ipc2581/src/types/ecad.rs
@@ -111,7 +111,8 @@ pub struct StackupLayer {
     pub spec_refs: Vec<Symbol>,
     pub dielectric_constant: Option<f64>,
     pub loss_tangent: Option<f64>,
-    pub layer_number: Option<u32>,
+    /// IPC sequence is a nonnegative double, not an integer layer ordinal.
+    pub layer_number: Option<f64>,
 }
 
 /// Step represents a design (board, panel, etc.)
